update timeline docstring
